# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Materials Management API with Golang",
        "title": "Materials Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bao Tran",
            "url": "http://github.com/remiehneppo",
            "email": "bao.tran080898@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api/v1",
    "paths": {
...
        "/materials-profiles": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve materials profiles based on filter criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials-profiles"
                ],
                "summary": "Filter materials profiles",
                "parameters": [
                    {
                        "description": "Materials profile filter",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.MaterialsProfileFilterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materials profiles retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.MaterialsProfile"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-profiles/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new materials profile for a specific maintenance instance and equipment machinery",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials-profiles"
                ],
                "summary": "Create a new materials profile",
                "parameters": [
                    {
                        "description": "Materials profile creation request",
                        "name": "materials_profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateMaterialProfileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materials profile created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-profiles/paginated": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve materials profiles with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials-profiles"
                ],
                "summary": "Get paginated materials profiles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated materials profiles retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-profiles/upload-estimate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an Excel sheet to update materials estimate profile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials-profiles"
                ],
                "summary": "Update materials estimate profile by uploading a sheet",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON request data containing project, maintenance_tier, maintenance_number, sheet_name, sector",
                        "name": "request",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materials estimate profile updated successfully",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-profiles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific materials profile using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials-profiles"
                ],
                "summary": "Get materials profile by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Materials Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materials profile retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.MaterialsProfile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Materials profile not found",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/materials-profiles/{id}/materials": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials-profiles"
                ],
                "summary": "Persist a material in Material Profile Estimate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Material Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Material",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/materialprofile.UpsertMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "materialprofile.UpsertMaterialRequest": {
            "type": "object",
            "required": [
                "category",
                "material"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "material": {
                    "$ref": "#/definitions/types.Material"
                }
            }
        },
...
        "types.CreateMaterialProfileReq": {
            "type": "object",
            "required": [
                "equipment_machinery_id",
                "estimate",
                "index_path",
                "maintenance_instance_id",
                "sector"
            ],
            "properties": {
                "equipment_machinery_id": {
                    "type": "string"
                },
                "estimate": {
                    "$ref": "#/definitions/types.MaterialsForEquipment"
                },
                "index_path": {
                    "type": "string"
                },
                "maintenance_instance_id": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                }
            }
        },
...
        "types.Material": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "types.MaterialsForEquipment": {
            "type": "object",
            "properties": {
                "consumable_supplies": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.Material"
                    }
                },
                "replacement_materials": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/types.Material"
                    }
                }
            }
        },
        "types.MaterialsProfile": {
            "type": "object",
            "properties": {
                "equipment_machinery_id": {
                    "type": "string"
                },
                "estimate": {
                    "$ref": "#/definitions/types.MaterialsForEquipment"
                },
                "id": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "maintenance_instance_id": {
                    "type": "string"
                },
                "reality": {
                    "$ref": "#/definitions/types.MaterialsForEquipment"
                },
                "sector": {
                    "type": "string"
                }
            }
        },
        "types.MaterialsProfileFilterRequest": {
            "type": "object",
            "properties": {
                "equipment_machinery_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "maintenance_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sector": {
                    "type": "string"
                }
            }
        },
        "types.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}